requirements.txt
# Microsoft Graph SDK
msgraph-sdk>=1.0.0
azure-identity>=1.12.0

# Authentication
msal>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Excel export
openpyxl>=3.1.0
pandas>=2.0.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Azure Speech Service (for audio transcription)
azure-cognitiveservices-speech>=1.32.0
